:building_construction: map + concatenation based FizzBuzz solution
i.e., the pretentious solution...

In all seriousness, the "naive" if-elseif solution:
1. runs faster
   - map iterator overhead
   - string concatenation takes more time than simply referencing a literal string
   - extra `str.empty()` check
2. uses less memory
   - map and map iterator overhead
3. is more readable
   - you're lying if your first thought at a solution was this
4. doesn't try to be "clever"
   - related to 3, the more "clever" your solution, the less people want
     to look at it. This map-strconcat is overkill for the simple FizzBuzz prompt.

With that said, part of the point is a data-driven coding architecture,
where the business logic doesn't have to change if you change certain
parameters, like the prime number set of divisors. This coding
architecture is also DRYer.